Add Faster version of Remove Element
Note: This version, two pointers, one moving from the head to the tail,
the other moving from the tail to the head, it is a little bit faster
than the previous version (in which, both pointer moving from the tail
to the head). It is a little bit trickier to handle the corner cases
though.